fix(persistence): split repository error contracts
