+++ catalog/persistence/src/repositories/category-entity-not-found.error.ts
@@ -1,0 +1,6 @@
+export class CategoryEntityNotFoundError extends Error {
+  constructor(id: string) {
+    super(`CategoryEntity with id ${id} was not found`)
+    this.name = 'CategoryEntityNotFoundError'
+  }
+}
