+++ catalog/persistence/src/repositories/category-group-entity-not-found.error.ts
@@ -1,10 +1,3 @@
-export class CategoryEntityNotFoundError extends Error {
-  constructor(id: string) {
-    super(`CategoryEntity with id ${id} was not found`)
-    this.name = 'CategoryEntityNotFoundError'
-  }
-}
-
 export class CategoryGroupEntityNotFoundError extends Error {
   constructor(id: string) {
     super(`CategoryGroupEntity with id ${id} was not found`)
