+++ collaboration/persistence/src/repositories/chat-entity-not-found.error.ts
@@ -1,0 +1,6 @@
+export class ChatEntityNotFoundError extends Error {
+  constructor(id: string) {
+    super(`ChatEntity with id ${id} was not found`)
+    this.name = 'ChatEntityNotFoundError'
+  }
+}
