+++ collaboration/persistence/src/repositories/discussion-entity-not-found.error.ts
@@ -1,0 +1,6 @@
+export class DiscussionEntityNotFoundError extends Error {
+  constructor(id: string) {
+    super(`DiscussionEntity with id ${id} was not found`)
+    this.name = 'DiscussionEntityNotFoundError'
+  }
+}
