+++ collaboration/persistence/src/repositories/message-entity-not-found.error.ts
@@ -1,0 +1,6 @@
+export class MessageEntityNotFoundError extends Error {
+  constructor(id: string) {
+    super(`MessageEntity with id ${id} was not found`)
+    this.name = 'MessageEntityNotFoundError'
+  }
+}
