+++ collaboration/persistence/src/repositories/project-entity-not-found.error.ts
@@ -1,0 +1,6 @@
+export class ProjectEntityNotFoundError extends Error {
+  constructor(id: string) {
+    super(`ProjectEntity with id ${id} was not found`)
+    this.name = 'ProjectEntityNotFoundError'
+  }
+}
