+++ collaboration/persistence/src/repositories/reply-entity-not-found.error.ts
@@ -1,0 +1,6 @@
+export class ReplyEntityNotFoundError extends Error {
+  constructor(id: string) {
+    super(`ReplyEntity with id ${id} was not found`)
+    this.name = 'ReplyEntityNotFoundError'
+  }
+}
