+++ collaboration/persistence/src/repositories/review-entity-not-found.error.ts
@@ -1,0 +1,6 @@
+export class ReviewEntityNotFoundError extends Error {
+  constructor(id: string) {
+    super(`ReviewEntity with id ${id} was not found`)
+    this.name = 'ReviewEntityNotFoundError'
+  }
+}
