+++ files/persistence/src/repositories/file-entity-not-found.error.ts
@@ -1,0 +1,6 @@
+export class FileEntityNotFoundError extends Error {
+  constructor(id: string) {
+    super(`FileEntity with id ${id} was not found`)
+    this.name = 'FileEntityNotFoundError'
+  }
+}
