+++ files/persistence/src/repositories/upload-entity-not-found.error.ts
@@ -1,10 +1,3 @@
-export class FileEntityNotFoundError extends Error {
-  constructor(id: string) {
-    super(`FileEntity with id ${id} was not found`)
-    this.name = 'FileEntityNotFoundError'
-  }
-}
-
 export class UploadEntityNotFoundError extends Error {
   constructor(id: string) {
     super(`UploadEntity with id ${id} was not found`)
