+++ portfolio/persistence/src/repositories/portfolio-entity-not-found.error.ts
